requirements.txt
pynput
pyaudio
pyyaml
PyGObject
